# docs/swagger/swagger.yaml
...
openapi: 3.0.1
info:
  title: FootballCheck
  version: '1.0'
  description: The API for the FootballCheck project
paths:
  "/api/v1/agents":
    get:
      tags:
      - Agent
      parameters:
      - name: sortState
        in: query
        schema:
          "$ref": "#/components/schemas/AgentSortState"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/AgentDto"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/AgentDto"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/AgentDto"
    post:
      security:
      - Bearer: []
      tags:
      - Agent
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AgentBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/AgentBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/AgentBaseDto"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/AgentDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/AgentDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/AgentDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/agents/{id}":
    put:
      security:
      - Bearer: []
      tags:
      - Agent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/AgentBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/AgentBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/AgentBaseDto"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/AgentDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/AgentDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/AgentDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      security:
      - Bearer: []
      tags:
      - Agent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/AgentDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/AgentDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/AgentDto"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - Agent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/AgentDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/AgentDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/AgentDto"
        '404':
          description: Not Found
  "/api/v1/clubs":
    get:
      tags:
      - Club
      parameters:
      - name: Name
        in: query
        schema:
          type: string
      - name: Country
        in: query
        schema:
          type: string
      - name: MinFoundationDate
        in: query
        schema:
          type: integer
          format: int32
      - name: MaxFoundationDate
        in: query
        schema:
          type: integer
          format: int32
      - name: sortState
        in: query
        schema:
          "$ref": "#/components/schemas/ClubSortState"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Club"
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Club"
            text/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Club"
    post:
      security:
      - Bearer: []
      tags:
      - Club
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ClubBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/ClubBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ClubBaseDto"
      responses:
        '201':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ClubDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/ClubDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/ClubDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '409':
          description: Conflict
  "/api/v1/clubs/{id}":
    put:
      security:
      - Bearer: []
      tags:
      - Club
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ClubBaseDto"
          text/json:
            schema:
              "$ref": "#/components/schemas/ClubBaseDto"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ClubBaseDto"
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ClubDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/ClubDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/ClubDto"
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '404':
          description: Not Found
        '409':
          description: Conflict
    delete:
      security:
      - Bearer: []
      tags:
      - Club
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ClubDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/ClubDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/ClubDto"
        '401':
          description: Unauthorized
        '404':
          description: Not Found
    get:
      tags:
      - Club
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                "$ref": "#/components/schemas/ClubDto"
            application/json:
              schema:
                "$ref": "#/components/schemas/ClubDto"
            text/json:
              schema:
                "$ref": "#/components/schemas/ClubDto"
        '404':
          description: Not Found
...
components:
  schemas:
    AgentBaseDto:
      type: object
      properties:
        playerId:
          type: integer
          format: int32
          nullable: true
        surname:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
      additionalProperties: false
    AgentDto:
      type: object
      properties:
        playerId:
          type: integer
          format: int32
          nullable: true
        surname:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
    AgentSortState:
      enum:
      - IdAsc
      - IdDesc
      - PlayerSurnameAsc
      - PlayerSurnameDesc
      - SurnameAsc
      - SurnameDesc
      - CountryAsc
      - CountryDesc
      type: string
    Club:
      required:
      - country
      - foundationDate
      - name
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        country:
          type: string
        foundationDate:
          type: integer
          format: int32
      additionalProperties: false
    ClubBaseDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        foundationDate:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    ClubDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        country:
          type: string
          nullable: true
        foundationDate:
          type: integer
          format: int32
          nullable: true
        id:
          type: integer
          format: int32
      additionalProperties: false
    ClubSortState:
      enum:
      - IdAsc
      - IdDesc
      - NameAsc
      - NameDesc
      - CountryAsc
      - CountryDesc
      - FoundationDateAsc
      - FoundationDateDesc
      type: string
...
  securitySchemes:
    Bearer:
      type: apiKey
      description: Please insert JWT with Bearer into field
      name: Authorization
      in: header